Listwise Deletion (LW)

This function deletes examinees who report missing responses (see De Ayala et al. 2001) <doi:10.1111/j.1745-3984.2001.tb01124.x>.

Usage

Listwise(test.data, Mvalue = "NA")

Arguments

test.data

Test data set (a data frame or a matrix) containing missing responses. Missing values are coded as NA or other values (e.g., 8, 9).

Mvalue

Missing response indicators in the data (e.g. "NA", "8", "9", etc.). Mvalue="NA" by default.

Value

A data frame with no missing responses.

References

De Ayala, R. J., Plake, B. S., & Impara, J. C. (2001). "The impact of omitted responses on the accuracy of ability estimation in item response theory." Journal of Educational Measurement, 38(3), 213–234. doi:10.1111/j.1745-3984.2001.tb01124.x.

Examples

 
        Listwise(test.data, Mvalue="NA")
#>     Item_1 Item_2 Item_3 Item_4 Item_5 Item_6 Item_7 Item_8 Item_9 Item_10
#> 14       1      1      1      1      1      1      1      0      0       0
#> 28       1      0      0      0      0      0      0      0      0       1
#> 32       1      0      0      1      1      0      0      1      1       0
#> 34       0      0      0      1      0      0      1      0      0       1
#> 38       1      0      0      0      1      0      0      1      0       0
#> 80       0      0      1      0      0      0      0      0      0       0
#> 85       1      1      0      1      1      1      1      1      1       1
#> 88       0      0      0      1      1      0      0      0      0       0
#> 108      0      0      1      1      1      0      0      1      0       0
#> 110      1      0      0      1      0      0      0      0      0       0
#> 115      1      1      1      1      1      0      1      1      1       1
#> 122      0      0      0      0      1      0      0      1      0       0
#> 126      0      0      1      1      1      0      0      0      1       1
#> 143      1      1      1      1      1      1      1      1      1       1
#> 180      0      0      1      0      1      0      0      1      0       0
#> 194      0      1      0      0      0      0      1      0      0       0
#> 202      1      1      0      0      0      1      0      0      0       0
#> 216      1      0      0      0      0      0      0      0      0       0
#> 218      1      0      0      0      0      0      1      0      0       0
#> 220      0      1      1      1      0      1      0      1      0       0
#> 223      1      0      1      0      1      0      1      1      0       1
#> 248      0      0      0      0      1      0      1      0      0       0
#> 254      0      1      0      0      1      0      0      0      0       0
#> 260      1      1      1      0      1      0      0      1      0       0
#> 262      0      0      0      1      1      0      0      0      1       0
#> 283      1      1      1      1      1      0      1      1      1       1
#> 284      1      1      0      1      0      1      1      0      0       0
#> 296      0      1      0      1      1      0      0      0      0       1
#> 300      0      0      0      1      0      0      0      0      0       0
#> 301      0      1      1      1      1      0      1      0      1       1
#> 304      1      0      1      1      0      0      0      0      0       1
#> 306      0      1      0      0      1      0      0      0      0       0
#> 350      0      0      1      0      1      0      0      0      0       1
#> 376      0      1      0      0      0      0      0      0      0       0
#> 380      1      1      0      0      0      1      0      0      0       0
#> 382      0      1      0      1      1      0      0      0      1       0
#> 384      0      0      0      0      0      0      0      0      0       0
#> 394      0      0      0      0      0      0      0      0      0       0
#> 400      0      0      0      0      0      0      1      0      0       0
#> 408      0      1      1      0      0      0      1      0      0       0
#> 412      0      1      0      0      0      0      0      0      0       0
#> 424      0      0      0      0      0      0      0      0      0       0
#> 428      0      0      0      0      1      0      0      0      0       0
#> 440      0      0      1      0      0      0      0      0      0       0
#> 442      1      1      0      1      1      1      1      1      0       0
#> 444      0      1      0      0      1      0      1      0      0       0
#> 446      1      0      0      0      1      0      0      1      0       0
#> 452      1      1      0      0      1      1      0      1      0       0
#> 454      0      1      0      1      0      0      1      0      0       0
#> 458      0      0      1      0      1      0      0      0      0       0
#> 462      1      1      1      0      0      1      1      0      0       0
#> 466      0      1      0      1      1      0      0      0      1       0
#> 483      1      1      0      1      1      1      1      1      1       1
#> 500      0      1      0      1      1      1      1      0      0       0
#> 520      0      1      0      0      0      0      0      0      0       0
#> 534      1      0      1      0      0      0      0      0      0       0
#> 539      0      0      1      1      1      0      0      1      0       1
#> 544      0      0      0      1      0      0      0      1      0       0
#> 554      0      1      1      0      1      0      0      0      0       0
#> 560      1      1      0      1      1      1      0      0      0       0
#> 567      1      1      0      1      1      1      1      1      1       1
#> 568      0      0      1      0      1      0      0      0      0       0
#> 592      0      0      1      1      0      0      0      0      0       0
#> 594      0      0      1      0      0      0      0      0      0       0
#> 596      0      1      1      1      0      0      0      0      0       0
#> 608      1      0      1      0      0      0      0      1      0       0
#> 616      0      0      0      0      0      0      0      0      0       0
#> 628      0      1      1      0      0      0      0      0      0       0
#> 632      0      1      0      1      0      0      0      0      0       0
#> 634      0      0      1      0      1      0      0      1      0       0
#> 648      0      0      0      0      0      0      0      0      0       0
#> 664      0      1      0      1      1      0      1      1      1       0
#> 665      1      1      0      1      1      1      1      1      1       1
#> 668      0      1      1      0      1      0      0      1      0       0
#> 672      1      1      0      1      0      0      0      1      0       0
#> 678      1      0      0      1      1      0      0      1      1       0
#> 682      0      0      1      0      1      0      0      0      0       0
#> 702      1      1      0      1      0      1      1      0      0       1
#> 708      1      0      0      0      0      0      0      1      0       1
#> 710      1      0      0      0      1      0      1      0      0       0
#> 720      0      0      1      0      0      0      0      0      0       1
#> 732      0      0      0      0      0      0      0      0      0       1
#> 741      1      1      1      1      1      1      1      1      1       1
#> 747      1      0      1      0      1      0      0      0      0       1
#> 760      0      0      1      0      0      0      0      0      0       1
#>     Item_11 Item_12 Item_13 Item_14 Item_15 Item_16 Item_17 Item_18 Item_19
#> 14        0       0       0       0       0       0       0       0       0
#> 28        0       0       0       0       0       0       0       0       0
#> 32        0       0       0       0       0       0       0       0       0
#> 34        0       0       0       0       1       0       0       0       0
#> 38        0       0       0       0       0       0       0       0       0
#> 80        0       0       0       0       0       0       0       0       0
#> 85        1       1       1       0       1       1       1       1       1
#> 88        0       0       0       1       0       0       0       0       0
#> 108       0       0       0       0       0       0       0       1       0
#> 110       1       0       0       0       0       0       1       0       0
#> 115       1       1       0       1       1       1       1       1       1
#> 122       1       0       0       1       1       0       0       0       0
#> 126       0       0       0       1       0       1       0       0       0
#> 143       1       1       1       1       1       1       1       1       1
#> 180       0       0       0       1       0       0       0       0       0
#> 194       0       0       0       0       0       0       0       0       0
#> 202       0       0       1       0       0       0       0       0       0
#> 216       0       0       0       0       0       0       0       0       0
#> 218       0       0       0       0       0       0       0       0       0
#> 220       0       0       0       0       1       0       0       0       0
#> 223       0       1       1       1       0       0       1       0       1
#> 248       0       0       0       0       0       0       0       0       0
#> 254       0       0       0       0       0       0       0       0       0
#> 260       0       1       0       0       0       0       0       0       0
#> 262       0       0       0       0       0       1       0       0       0
#> 283       1       1       1       1       1       1       1       1       1
#> 284       0       1       0       1       0       0       1       0       0
#> 296       0       0       0       0       0       0       0       0       0
#> 300       0       0       0       0       0       0       0       0       0
#> 301       1       0       1       1       0       0       1       0       0
#> 304       0       0       0       0       1       0       0       0       0
#> 306       0       0       0       0       0       0       0       0       0
#> 350       0       0       0       0       0       0       0       0       0
#> 376       0       0       0       0       0       0       0       0       0
#> 380       1       0       0       0       1       0       0       0       0
#> 382       0       0       0       0       1       0       0       0       0
#> 384       1       0       0       0       0       0       0       0       0
#> 394       0       0       0       0       0       0       0       0       0
#> 400       0       0       0       0       0       0       1       0       0
#> 408       1       0       0       0       1       1       0       0       0
#> 412       0       0       1       0       0       0       0       0       0
#> 424       0       0       0       0       0       0       0       0       1
#> 428       0       0       0       0       0       0       0       0       0
#> 440       0       0       1       0       0       0       0       0       0
#> 442       0       0       0       0       0       0       0       1       0
#> 444       0       0       0       0       0       0       0       0       0
#> 446       1       0       0       0       0       0       0       0       0
#> 452       0       0       0       0       0       0       0       0       0
#> 454       0       0       0       0       0       0       0       0       0
#> 458       0       0       0       0       0       0       0       0       0
#> 462       0       0       0       1       1       0       0       0       0
#> 466       0       0       0       0       0       0       0       0       1
#> 483       0       0       1       1       0       1       0       1       0
#> 500       0       0       1       0       0       0       0       0       0
#> 520       0       0       0       0       0       0       0       0       0
#> 534       0       0       0       0       0       0       0       0       0
#> 539       0       0       0       1       0       0       0       0       1
#> 544       0       0       0       0       0       0       1       0       1
#> 554       0       0       0       0       0       0       1       0       0
#> 560       0       0       0       0       0       0       0       0       0
#> 567       1       1       1       1       1       1       1       1       1
#> 568       0       0       0       0       0       0       0       0       0
#> 592       0       1       0       0       0       0       0       0       0
#> 594       0       0       0       0       0       0       0       1       1
#> 596       0       0       0       0       0       1       0       0       0
#> 608       0       0       0       0       0       0       1       0       0
#> 616       0       0       0       0       0       1       0       0       0
#> 628       0       0       0       0       0       1       0       0       0
#> 632       0       0       0       0       1       0       0       0       0
#> 634       1       0       0       1       0       0       1       1       0
#> 648       0       0       0       0       0       0       0       0       0
#> 664       0       0       0       0       0       0       0       0       0
#> 665       1       1       1       1       1       1       1       1       1
#> 668       1       0       1       0       0       1       0       0       0
#> 672       0       0       0       0       0       0       1       0       0
#> 678       0       0       0       0       0       0       0       0       0
#> 682       0       0       0       0       0       0       0       0       0
#> 702       0       0       0       1       0       0       0       1       0
#> 708       0       0       0       0       0       0       0       0       0
#> 710       0       0       0       0       0       0       0       0       1
#> 720       0       0       0       0       0       0       0       0       0
#> 732       0       0       1       0       0       0       0       0       0
#> 741       1       1       1       1       1       1       1       1       1
#> 747       0       1       0       0       0       0       1       1       0
#> 760       1       0       0       0       0       0       0       0       0
#>     Item_20
#> 14        0
#> 28        0
#> 32        0
#> 34        0
#> 38        0
#> 80        0
#> 85        1
#> 88        0
#> 108       1
#> 110       1
#> 115       0
#> 122       0
#> 126       0
#> 143       1
#> 180       0
#> 194       0
#> 202       0
#> 216       0
#> 218       1
#> 220       0
#> 223       1
#> 248       0
#> 254       0
#> 260       0
#> 262       0
#> 283       1
#> 284       0
#> 296       0
#> 300       0
#> 301       1
#> 304       0
#> 306       0
#> 350       0
#> 376       0
#> 380       0
#> 382       0
#> 384       0
#> 394       0
#> 400       0
#> 408       0
#> 412       0
#> 424       0
#> 428       0
#> 440       0
#> 442       0
#> 444       0
#> 446       0
#> 452       0
#> 454       0
#> 458       0
#> 462       1
#> 466       0
#> 483       0
#> 500       1
#> 520       0
#> 534       0
#> 539       0
#> 544       1
#> 554       0
#> 560       1
#> 567       1
#> 568       0
#> 592       0
#> 594       1
#> 596       0
#> 608       0
#> 616       0
#> 628       0
#> 632       0
#> 634       0
#> 648       0
#> 664       0
#> 665       1
#> 668       1
#> 672       0
#> 678       0
#> 682       1
#> 702       0
#> 708       0
#> 710       0
#> 720       0
#> 732       0
#> 741       1
#> 747       1
#> 760       0